containers/policy-engine/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pyyaml>=6.0
cryptography>=41.0.0
